removed refs for scroll, added HashLinks
