+++ package.json
@@ -27,6 +27,7 @@
     "react-hook-form": "^7.30.0",
     "react-redux": "^8.0.1",
     "react-router-dom": "^6.3.0",
+    "react-router-hash-link": "^2.4.3",
     "react-scripts": "5.0.1",
     "react-scroll": "^1.8.7",
     "react-slick": "^0.29.0",
