+++ src/components/Calendar/CalendarGrid/EventsBlocks.tsx
@@ -1,0 +1,12 @@
+import { CalendarEvent } from '../CalendarStyled'
+
+const EventsBlocks = () => {
+  return (
+    <div>
+      <CalendarEvent> some event</CalendarEvent>
+      <CalendarEvent> some event 1</CalendarEvent>
+    </div>
+  )
+}
+
+export default EventsBlocks
