+++ src/components/Calendar/CalendarGrid/EventsList.tsx
@@ -1,0 +1,12 @@
+import { CalendarEvent, EventsNumber } from '../CalendarStyled'
+
+const EventsList = () => {
+  return (
+    <div>
+      <CalendarEvent> EVENTS</CalendarEvent>
+      <EventsNumber>4</EventsNumber>
+    </div>
+  )
+}
+
+export default EventsList
