--- src/components/Calendar/Header/Header.tsx
@@ -1,5 +1,0 @@
-const Header = () => {
-  return <div></div>
-}
-
-export { Header }
