+++ src/components/HomePage/UpcomingEvents/UpcomingEvents.tsx
@@ -1,12 +1,11 @@
-import { forwardRef, RefObject } from 'react'
 import { Divider } from '@mui/material'
 import useStyles from '../../styles/useStyle'
 import ScrollableItems from '../ScrollableItems/ScrollableItems'
 
-const UpcomingEvents = forwardRef(({ upcomingRef }: { upcomingRef: RefObject<HTMLDivElement> }) => {
+const UpcomingEvents: React.FC = () => {
   const classes = useStyles()
   return (
-    <div ref={upcomingRef} className={classes.homeBlocks}>
+    <div id="upcoming-events" className={classes.homeBlocks}>
       <div className={classes.blockTitle}>
         <div id="title">Upcoming events</div>
         <div id="view-all">View all</div>
@@ -15,6 +14,6 @@ const UpcomingEvents = forwardRef(({ upcomingRef }: { upcomingRef: RefObject<HTM
       <ScrollableItems />
     </div>
   )
-})
+}
 
 export default UpcomingEvents
