+++ src/components/NavBar/UserBarLoggedOut.tsx
@@ -1,9 +1,9 @@
-import useStyles from '../../styles/useStyle'
+import useNavbarStyles from './useNavbarStyles'
 import clsx from 'clsx'
 import { Avatar } from '@mui/material'
 
 const UserBarLoggedOut = () => {
-  const classes = useStyles()
+  const classes = useNavbarStyles()
   return (
     <div className={clsx(classes.userBarInner, classes.flexCenter)}>
       <Avatar sx={{ width: 30, height: 30 }} src="/broken-image.jpg" />
