+++ src/components/NavBar/userBar.tsx
@@ -1,14 +1,14 @@
-import useStyles from '../../styles/useStyle'
+import useNavbarStyles from './useNavbarStyles'
 import ArrowButton from './UserMenu/arrowButton'
 import UserBarLoggedIn from './UserBarLoggedIn'
 import UserBarLoggedOut from './UserBarLoggedOut'
 import clsx from 'clsx'
 import { memo } from 'react'
-import { RootState } from '../../../store/index'
+import { RootState } from '../../store/index'
 import { useSelector } from 'react-redux'
 
 const UserBar: React.FC = () => {
-  const classes = useStyles()
+  const classes = useNavbarStyles()
   const isUserAuth = useSelector<RootState, boolean>(state => state.ui.isUserAuth)
 
   return (
